fixed a bug

+++ src/cpp/Solver.cpp
@@ -28,8 +28,7 @@ template <class type> void clear(type* a, unsigned int N)
 
 CSolver::CSolver(CSkylineMatrix<double>* K) : K(K) {};
 
-CSolver::CSolver(CSkylineMatrix<double>* K, CSkylineMatrix<double>* M) : K(K), M(M) {
-};	
+CSolver::CSolver(CSkylineMatrix<double>* K, CSkylineMatrix<double>* M) : K(K), M(M) {};	
 
 
 // Calculate the lumped mass matrix
@@ -400,7 +399,7 @@ void CG_alpha::G_alpha_Intregration(CLoadCaseData& Load, int i_load)
 		His_Output->OutputHisMessage(t, dis, vel, acc, N_His_Freedom, num_freedom);
 
 		// Tecplot Output
-		if (fmod(Tec_Count, Ani_Interval) == 0) {
+		if (fmod((double)Tec_Count, (double)Ani_Interval) == 0) {
 			cout << "Output Tecplot and Paraview, Time =  " << t << endl;
 			Tecplot_Output->OutputTecplot(t, dis);
 			Paraview_Output->OutputVTK(t, dis);
